+++ src/components/MovieInformation.js
@@ -3,7 +3,7 @@ import propTypes from "prop-types";
 import { Link } from "react-router-dom";
 import FontAwesomeIcon from "@fortawesome/react-fontawesome";
 import PosterCard from "./PosterCard";
-import { normalizeMovie } from "../api/APIUtils";
+import { normalizeMovie } from "../api/tmdb";
 import Scroll from "./Scroll";
 import Trailers from "./Trailers";
 import "../css/MovieInfo.scss";
