+++ src/components/WatchList/CardList.js
@@ -3,7 +3,7 @@ import PropTypes from "prop-types";
 import FontAwesomeIcon from "@fortawesome/react-fontawesome";
 import moment from "moment-mini";
 import { Link } from "react-router-dom";
-import { getFullImgPath } from "../../api/APIUtils";
+import { getFullImgPath } from "../../api/tmdb";
 import ListDeleteBtn from "./ListDeleteBtn";
 import ListMoveBtn from "./ListMoveBtn";
 import LoadingCardList from "../Loading/LoadingCardList";
